+++ ktrain/text/ner/anago/models.py
@@ -3,10 +3,10 @@ Model definition.
 """
 from ....imports import *
 from .... import utils as U
-if U.is_tf_keras():
-    from .layers import CRF
-else:
-    from .layers_standalone import CRF
+#if U.is_tf_keras():
+    #from .layers import CRF
+#else:
+    #from .layers_standalone import CRF
 
 
 def save_model(model, weights_file, params_file):
@@ -46,7 +46,9 @@ class BiLSTMCRF(object):
                  dropout=0.5,
                  embeddings=None,
                  use_char=True,
-                 use_crf=True):
+                 use_crf=True,
+                 mask_zero=True,
+                 use_elmo=False):
         """Build a Bi-LSTM CRF model.
 
         Args:
@@ -62,6 +64,9 @@ class BiLSTMCRF(object):
             embeddings (numpy array): word embedding matrix.
             use_char (boolean): add char feature.
             use_crf (boolean): use crf as last layer.
+            mask_zero(boolean): mask zero
+            use_elmo(boolean): If True, model will be configured to accept Elmo embeddings
+                               as an additional input to word and character embeddings
         """
         super(BiLSTMCRF).__init__()
         self._char_embedding_dim = char_embedding_dim
@@ -76,22 +81,31 @@ class BiLSTMCRF(object):
         self._use_crf = use_crf
         self._embeddings = embeddings
         self._num_labels = num_labels
+        self.mask_zero = mask_zero
+        self._use_elmo = use_elmo
+        # NOTE: provide option for mask_zero=False because TF2 with V2 behavior throws error otherwise
+        # REFERENCE:
+        #   https://github.com/tensorflow/tensorflow/issues/33069
+        #   https://github.com/tensorflow/tensorflow/issues/33148
 
     def build(self):
+
         # build word embedding
         word_ids = Input(batch_shape=(None, None), dtype='int32', name='word_input')
         inputs = [word_ids]
+        embedding_list = []
         if self._embeddings is None:
             word_embeddings = Embedding(input_dim=self._word_vocab_size,
                                         output_dim=self._word_embedding_dim,
-                                        mask_zero=True,
+                                        mask_zero=self.mask_zero,
                                         name='word_embedding')(word_ids)
         else:
             word_embeddings = Embedding(input_dim=self._embeddings.shape[0],
                                         output_dim=self._embeddings.shape[1],
-                                        mask_zero=True,
+                                        mask_zero=self.mask_zero,
                                         weights=[self._embeddings],
                                         name='word_embedding')(word_ids)
+        embedding_list.append(word_embeddings)
 
         # build character based word embedding
         if self._use_char:
@@ -99,16 +113,22 @@ class BiLSTMCRF(object):
             inputs.append(char_ids)
             char_embeddings = Embedding(input_dim=self._char_vocab_size,
                                         output_dim=self._char_embedding_dim,
-                                        mask_zero=True,
+                                        mask_zero=self.mask_zero,
                                         name='char_embedding')(char_ids)
             char_embeddings = TimeDistributed(Bidirectional(LSTM(self._char_lstm_size)))(char_embeddings)
-            word_embeddings = Concatenate()([word_embeddings, char_embeddings])
+            embedding_list.append(char_embeddings)
+        if self._use_elmo:
+            elmo_embeddings = Input(shape=(None, 1024), dtype='float32')
+            inputs.append(elmo_embeddings)
+            embedding_list.append(elmo_embeddings)
+        word_embeddings = Concatenate()(embedding_list) if len(embedding_list) > 1 else embedding_list[0]
 
         word_embeddings = Dropout(self._dropout)(word_embeddings)
         z = Bidirectional(LSTM(units=self._word_lstm_size, return_sequences=True))(word_embeddings)
         z = Dense(self._fc_dim, activation='tanh')(z)
 
         if self._use_crf:
+            from .layers import CRF
             crf = CRF(self._num_labels, sparse_target=False)
             loss = crf.loss_function
             pred = crf(z)
@@ -119,3 +139,4 @@ class BiLSTMCRF(object):
         model = Model(inputs=inputs, outputs=pred)
 
         return model, loss
+